Separated the question in their own components, the header too
